Safari-609.1.11.1/LayoutTests/editing/find/find-and-replace-at-editing-boundary-expected.txt

Verifies that find and replace ignores matches that span editing boundaries. This test requires WebKitTestRunner.

After replacing 'apple' with 'pear':
| "
        "
| <span>
|   contenteditable=""
|   "<#selection-anchor>pear<#selection-focus> ap"
| <span>
|   "ple apple"
| "
    "
